# storage/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Documentation for the Application API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Laravel Server #1"
        }
    ],
    "paths": {
...
        "/users/all": {
            "get": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "users"
                ]
            }
        },
...
    },
...
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}